--- expo/App.tsx
@@ -1,3 +1,0 @@
-import App from "./src/screens/Home";
-
-export default App;
